Support playback & seeking in media files over 2G.
BUG=95805
TEST=unittests of touched files pass, manually tested w/ sonyh2.webm which is >5G (internal only), and trybots.


Review URL: http://codereview.chromium.org/8349009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106183 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d8f372ad864e13f3f06bebec3a1189b8f8583ccc

+++ webkit/glue/media/buffered_resource_loader.cc
@@ -734,12 +734,12 @@ void BufferedResourceLoader::ReadInternal() {
 
 bool BufferedResourceLoader::VerifyPartialResponse(
     const WebURLResponse& response) {
-  int first_byte_position, last_byte_position, instance_size;
+  int64 first_byte_position, last_byte_position, instance_size;
 
   if (!MultipartResponseDelegate::ReadContentRanges(response,
-                         &first_byte_position,
-                         &last_byte_position,
-                         &instance_size)) {
+                                                    &first_byte_position,
+                                                    &last_byte_position,
+                                                    &instance_size)) {
     return false;
   }
