+++ webkit/glue/multipart_response_delegate.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2006-2009 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
@@ -91,9 +91,9 @@ class MultipartResponseDelegate {
   // in a multipart response.
   // Returns true on success.
   static bool ReadContentRanges(const WebKit::WebURLResponse& response,
-                                int* content_range_lower_bound,
-                                int* content_range_upper_bound,
-                                int* content_range_instance_size);
+                                int64* content_range_lower_bound,
+                                int64* content_range_upper_bound,
+                                int64* content_range_instance_size);
 
  private:
   friend class MultipartResponseDelegateTester;  // For unittests.
